+++ apps/api/exceptions.py
@@ -1,5 +1,0 @@
-class MySmileApiException(Exception):
-
-    def __init__(self, msg, code):
-        self.msg = msg
-        self.code = code