+++ apps/api/urls.py
@@ -1,10 +1,6 @@
-from django.conf.urls import patterns, url
-from apps.api.views import MySmileApi
-from django.views.decorators.csrf import csrf_exempt
-
+from django.conf.urls import *
 
 urlpatterns = patterns('',
-    url(r'^(?P<resource>[a-z,0-9]+)$', csrf_exempt(MySmileApi.as_view())),
+    #url(r'', include('apps.api.v2.urls', namespace='default')),
+    url(r'^v1/', include('apps.api.v1.urls', namespace='v1')),
 )
-
-