+++ apps/api/v1/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class ApiConfig(AppConfig):
-    name = 'apps.api'
+    name = 'apps.api.v1'
     verbose_name = 'Api'
     label = 'api'