+++ apps/utils/signals.py
@@ -29,11 +29,15 @@ def clear_cache(sender, instance=None, created=False, **kwargs):
     if sender.__name__=='Preferences':
         try:
             settings.MYSMILE_THEME = Preferences.objects.filter(key='THEME').values_list('value', flat=True)[0]
-
         except Exception as err:
             settings.MYSMILE_THEME = 'modern'
             logger.error(err)
 
+        try:
+            settings.MYSMILE_REST_API = 'True' in Preferences.objects.filter(key='REST_API').values_list('value', flat=True)
+        except Exception as err:
+            settings.MYSMILE_REST_API = False
+            logger.error(err)
 
 
 def clear_photo_file(sender, instance, **kwargs):