+++ mysmile/settings/local.py
@@ -7,6 +7,7 @@ from config.mysmile import *
 
 
 APP_MIDDLEWARE_CLASSES = (
+    'apps.api.middlewares.VersionSwitchMiddleware',
     'apps.utils.middlewares.ExceptionLoggingMiddleware',
     'apps.utils.middlewares.AdminLocaleOneLangMiddleware',
 )
@@ -15,11 +16,13 @@ THIRD_PARTY_MIDDLEWARE_CLASSES = (
     'debug_toolbar.middleware.DebugToolbarMiddleware',
     )
 
-MIDDLEWARE_CLASSES = DJANGO_MIDDLEWARE_CLASSES + APP_MIDDLEWARE_CLASSES + THIRD_PARTY_MIDDLEWARE_CLASSES
+MIDDLEWARE_CLASSES = DJANGO_MIDDLEWARE_CLASSES + \
+                     APP_MIDDLEWARE_CLASSES + \
+                     THIRD_PARTY_MIDDLEWARE_CLASSES
 
 # Apps specific for this project go here.
 LOCAL_APPS = (
-    'apps.api',
+    'apps.api.v1',
     'apps.pages',
     'apps.preferences',
     'apps.sitemap',
@@ -30,6 +33,7 @@ LOCAL_APPS = (
 # another apps
 THIRD_PARTY_APPS = ('debug_toolbar',
                     'compressor',
+                    'rest_framework',
                     )
 
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps