+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 
 
 **2015-09-## Version 0.8.0**
-  - 
+  - Refactor MySmile api on Django Rest Framework 3 
   -
 
 **2015-07-13 Version 0.7.0**